+++ wrapped-mc/src/main/java/com/mojang/blaze3d/systems/GpuDevice.java
@@ -9,5 +9,5 @@ public interface GpuDevice {
 
     CommandEncoder createCommandEncoder();
 
-    GpuBuffer createBuffer(@Nullable Supplier<String> var1, int var2, int var3);
+    GpuBuffer createBuffer(@Nullable Supplier<String> var1, int var2, long var3);
 }
